requirements.txt
fastapi
uvicorn
websockets
pyaudio
faster-whisper
google-generativeai
pyqt6
python-dotenv
numpy
